requirements.txt
pandas
pandas_ta
matplotlib
python-binance
